requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.8.1
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.0
idna==3.10
iopath==0.1.10
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.13
wheel==0.45.1
yacs==0.1.8
